+++ src/utils/styles/index.ts
@@ -1,4 +1,4 @@
-export { createStyles } from './createStyles'
+export { createStyles, createTypedStyles } from './createStyles'
 export { keyframes } from './keyframes'
 export { fontFace } from './fontFace'
 export { cssVariables, cssVar, injectCssVariables } from './cssVariables'